requirements.txt
modelscope==1.16.1
Pillow==11.0.0
datasets==2.21.0
datasketch==1.6.4
Flask==3.0.3
Flask_Cors==4.0.0
jieba==0.42.1
jsonlines==4.0.0
marshmallow==3.22.0
matplotlib==3.10.0
ngrok==1.4.0
nltk==3.8
numpy==1.26.4
pandas==2.2.3
pytz==2024.2
python-dateutil==2.9.0.post0
six==1.16.0
openai==1.59.6
requests==2.32.3
urllib3==2.2.3
charset-normalizer==3.4.0
idna==3.10
certifi==2024.8.30
peft==0.7.1
absl-py==2.1.0
psutil==5.9.8
pydantic==2.11.5
protobuf==4.25.3
packaging==24.1
rich==13.7.1
scikit_learn==1.5.1
joblib==1.4.2
scipy==1.11.4
threadpoolctl==3.5.0
sentence_transformers==2.3.1
huggingface_hub==0.34.0
tqdm==4.66.5
filelock==3.16.1
simhash==2.1.2
tiktoken==0.10.0
tokenizers==0.22.0
regex==2024.11.6
transformers==4.57.1
safetensors==0.4.5
jinja2==3.1.2
markupsafe==3.0.2
jsonlines==4.0.0
trl==0.13.0
ujson==5.11.0
wandb==0.18.3
streamlit==1.50.0
einops==0.8.1
swanlab==0.6.8
nvidia-ml-py==12.560.30
pyyaml==6.0.2
sympy==1.13.1
mpmath==1.3.0
torch==2.6.0
torchvision==0.21.0
tensorboard==2.17.1
typing_extensions>=4.5.0
